requirements.txt
pandas==2.0.3
SQLAlchemy==2.0.19
python-dotenv==1.0.0
pytest==7.4.0
pylint==2.17.5
flake8==6.1.0